Backend_fd/requirements.txt
cmake
wheel
face_recognition==1.4.0
fastapi==0.78.0
numpy==1.22.4
opencv_python==4.5.5.64
Pillow==9.1.1
starlette==0.14.2
python-multipart
uvicorn
